clean lockscreen settings and remove duplicate bottom shortcuts

+++ src/com/android/settings/rr/CustomHeader.java
@@ -105,7 +105,7 @@ public class CustomHeader extends SettingsPreferenceFragment  implements Prefere
         mHeaderShadow = (SeekBarPreference) findPreference(CUSTOM_HEADER_IMAGE_SHADOW);
         final int headerShadow = Settings.System.getInt(getContentResolver(),
                 Settings.System.STATUS_BAR_CUSTOM_HEADER_SHADOW, 0);
-        mHeaderShadow.setValue((int)((headerShadow / 255) * 100));
+        mHeaderShadow.setValue((int)(((double) headerShadow / 255) * 100));
         mHeaderShadow.setOnPreferenceChangeListener(this);
 
 }
